8286378: Address possibly lossy conversions in java.base Add explicit casts to replace implicit casts

+++ src/java.base/share/classes/java/io/ByteArrayInputStream.java
@@ -228,7 +228,7 @@ public class ByteArrayInputStream extends InputStream {
             k = n < 0 ? 0 : n;
         }
 
-        pos += k;
+        pos += (int)k;
         return k;
     }
